Added icons, god rays, SFX and dimming functionality to music zones
When any zone is activated for the first time in the session, the ZED image brightness is lowered to simulate dimming the studio lights.
Stepping into a zone turns on its spotlight (god ray sprite), particle system, and audio track.
Stepping out of a zone turns off all of those, as long as no one else is inside.

(Spotty implementation via a counter on TriggerEnter and TriggerExit. When tracking is interrupted and a person disappears, it doesn't count as a TriggerExit - so the zone considers them a new person when they reappear, and then the count is 1 too high.  I want to rewrite the script to loop through every visitor's position and double check that none of them are within the trigger bounds before deactivating, but it'll take up time I should spend on the keynote... so fingers crossed haha)

(...Could I check the ZED ID of the person to avoid duplicates? Hmmm)